[stories] update story name
