+++ stories/components/bulk-textarea/docs/BulkTextarea.stories.tsx
@@ -1,0 +1,40 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import BulkTextarea from '@semcore/bulk-textarea';
+
+import BasicUsageExample from './examples/basic-usage';
+
+const meta: Meta<typeof BulkTextarea> = {
+  title: 'Components/BulkTextarea/Documentation',
+  component: BulkTextarea,
+};
+
+export default meta;
+
+export const BasicUsage: StoryObj<typeof BulkTextarea> = {
+  args: {
+    ofRows: 30,
+    size: 'm',
+    rowsDelimiters: [','],
+    readonly: false,
+    disabled: false,
+    placeholder: 'Placeholder',
+    minRows: 2,
+    maxRows: 10,
+    validateOn: [
+      // 'blur',
+      'blurRow',
+    ],
+    pasteProps: {
+      delimiter: '\n',
+      skipEmptyRows: true,
+      rowProcessing: (row) => {
+        return row.replace(/http:\/\//, '');
+      },
+    },
+    rowProcessing: (row) => {
+      return row.replace(/http:\/\//, '');
+    },
+  },
+  render: BasicUsageExample,
+};
